requirements.txt
fastapi
uvicorn
yfinance
pandas
pydantic
python-dotenv
pytest
httpx